ripple-ml-local/requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
torch==2.4.1
torch-geometric==2.6.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1

# ── Database (replaces BigQuery) ───────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0

# ── Serving ────────────────────────────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.6.1

# ── Local LLM (replaces Gemini) ────────────────────────────────────────────────
ollama==0.3.3          # pip install ollama  — calls local Ollama server
httpx==0.27.2          # fallback HTTP calls

# ── Visualisation ──────────────────────────────────────────────────────────────
matplotlib==3.9.2
seaborn==0.13.2

# ── Utils ──────────────────────────────────────────────────────────────────────
tqdm==4.67.0
loguru==0.7.2
pyyaml==6.0.2
python-dotenv==1.0.1
typer==0.13.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0